refactor(components): aitextgenerate aitextevaluate listevaluate and listfilter to use the ai tool provider manager inestad of calling methods directly
