+++ src/SmartHopper.Core.Grasshopper/Tools/GhGetTools.cs
@@ -27,10 +27,8 @@ namespace SmartHopper.Core.Grasshopper.Tools
     /// <summary>
     /// Tool provider for Grasshopper component retrieval via AI Tool Manager.
     /// </summary>
-    public class GhGetTools : IAIToolProvider
+    public class GhGetTools
     {
-        #region Synonyms
-
         /// <summary>
         /// Synonyms for filter tags.
         /// Available filter tokens:
@@ -48,7 +46,7 @@ namespace SmartHopper.Core.Grasshopper.Tools
         ///   visible → previewon
         ///   hidden → previewoff.
         /// </summary>
-        private static readonly Dictionary<string, string> FilterSynonyms = new ()
+        public static readonly Dictionary<string, string> FilterSynonyms = new ()
         {
             { "LOCKED", "DISABLED" },
             { "UNLOCKED", "ENABLED" },
@@ -78,7 +76,7 @@ namespace SmartHopper.Core.Grasshopper.Tools
         ///   processingcomponents, intermediate, middle, middlecomponents → processing
         ///   isolatedcomponents → isolated.
         /// </summary>
-        private static readonly Dictionary<string, string> TypeSynonyms = new ()
+        public static readonly Dictionary<string, string> TypeSynonyms = new ()
         {
             { "PARAM", "PARAMS" },
             { "PARAMETER", "PARAMS" },
@@ -100,7 +98,7 @@ namespace SmartHopper.Core.Grasshopper.Tools
         /// Available Grasshopper component categories (e.g. Params, Maths, Vector, Curve, Surface, Mesh, etc.).
         /// Maps common abbreviations or alternate names to canonical category tokens.
         /// </summary>
-        private static readonly Dictionary<string, string> CategorySynonyms = new ()
+        public static readonly Dictionary<string, string> CategorySynonyms = new ()
         {
             { "PARAM", "PARAMS" },
             { "PARAMETERS", "PARAMS" },
@@ -125,9 +123,6 @@ namespace SmartHopper.Core.Grasshopper.Tools
             { "RHINOCEROS", "RHINO" },
             { "KANGAROOPHYSICS", "KANGAROO" },
         };
-        #endregion
-
-        #region HelperMethods
 
         /// <summary>
         /// Helper to parse include/exclude tokens.
@@ -135,7 +130,7 @@ namespace SmartHopper.Core.Grasshopper.Tools
         /// <param name="rawGroups">List of raw filter groups.</param>
         /// <param name="synonyms">Dictionary of synonyms for tokens.</param>
         /// <returns>Tuple of include and exclude sets.</returns>
-        private static (HashSet<string> Include, HashSet<string> Exclude) ParseIncludeExclude(IEnumerable<string> rawGroups, Dictionary<string, string> synonyms)
+        public static (HashSet<string> Include, HashSet<string> Exclude) ParseIncludeExclude(IEnumerable<string> rawGroups, Dictionary<string, string> synonyms)
         {
             var include = new HashSet<string>();
             var exclude = new HashSet<string>();
@@ -175,535 +170,5 @@ namespace SmartHopper.Core.Grasshopper.Tools
 
             return (include, exclude);
         }
-        #endregion
-
-        #region ToolRegistration
-
-        /// <summary>
-        /// Returns a list of AI tools provided by this plugin.
-        /// </summary>
-        /// <returns>Collection of AI tools.</returns>
-        public IEnumerable<AITool> GetTools()
-        {
-            yield return new AITool(
-                name: "gh_get",
-                description: "Read the current Grasshopper file with optional filters. By default, it returns all components. Returns a GhJSON structure of the file.",
-                parametersSchema: @"{
-                    ""type"": ""object"",
-                    ""properties"": {
-                        ""attrFilters"": {
-                            ""type"": ""array"",
-                            ""items"": { ""type"": ""string"" },
-                            ""description"": ""Optional array of attribute filter tokens. '+' includes, '-' excludes. Defaults to all components. Available tags:\n  selected/unselected: component selection state on canvas;\n  enabled/disabled: whether the component can run (enabled = unlocked);\n  error/warning/remark: runtime message levels;\n  previewcapable/notpreviewcapable: supports geometry preview;\n  previewon/previewoff: current preview toggle.\nSynonyms: locked→disabled, unlocked→enabled, remarks/info→remark, warn/warnings→warning, errors→error, visible→previewon, hidden→previewoff. Examples: '+error' → only components with errors; '+error +warning' → errors OR warnings; '+error -warning' → errors excluding warnings; '+error -previewoff' → errors with preview on; no filter → all components.""
-                        },
-                        ""typeFilter"": {
-                            ""type"": ""array"",
-                            ""items"": { ""type"": ""string"" },
-                            ""description"": ""Optional array of type tokens with include/exclude syntax. Defaults to all types. Available tokens:\n  params: only parameter objects;\n  components: only component objects;\n  input: components with no incoming connections;\n  output: components with no outgoing connections;\n  processing: components with both incoming and outgoing connections;\n  isolated: components with neither incoming nor outgoing connections.\nExamples: ['+params', '-components'] to include parameters and exclude components.""
-                        },
-                        ""guidFilter"": {
-                            ""type"": ""array"",
-                            ""items"": { ""type"": ""string"" },
-                            ""description"": ""Optional list of component GUIDs for initial filtering. When provided, only components with these GUIDs are processed. If not provided, all components are processed.""
-                        },
-                        ""connectionDepth"": {
-                            ""type"": ""integer"",
-                            ""default"": 0,
-                            ""description"": ""Depth of connections to include: 0 (default) only matching components; 1 includes directly connected components; 2 includes two-level connected components, etc.""
-                        }
-                    }
-                }",
-                execute: this.GhGetToolAsync
-            );
-
-            yield return new AITool(
-                name: "gh_get_available_components",
-                description: "Retrieve a list of all installed components in this current environment. Returns a JSON dictionary with names, GUIDs, categories, subcategories, descriptions, and keywords.",
-                parametersSchema: @"{
-                    ""type"": ""object"",
-                    ""properties"": {
-                        ""categoryFilter"": {
-                            ""type"": ""array"",
-                            ""items"": { ""type"": ""string"" },
-                            ""description"": ""Optionally filter components by category. '+' includes, '-' excludes. Most common categories: Params, Maths, Vector, Curve, Surface, Mesh, Intersect, Transform, Sets, Display, Rhino, Kangaroo. E.g. ['+Maths','-Params']. (note: use the tool 'gh_categories' to get the full list of available categories)""
-                        }
-                    }
-                }",
-                execute: this.GhRetrieveToolAsync
-            );
-
-            yield return new AITool(
-                name: "gh_categories",
-                description: "List all available Grasshopper component categories and subcategories with optional soft string filter.",
-                parametersSchema: @"{
-                    ""type"": ""object"",
-                    ""properties"": {
-                        ""filter"": {
-                            ""type"": ""string"",
-                            ""description"": ""Soft filter: return categories or subcategories containing the search tokens (split by space).""
-                        }
-                    }
-                }",
-                execute: this.GhCategoriesToolAsync
-            );
-        }
-        #endregion
-
-        #region GhGet
-
-        /// <summary>
-        /// Executes the Grasshopper get components tool.
-        /// </summary>
-        /// <param name="parameters">Parameters object containing filter settings.</param>
-        /// <returns>Task that returns the result of the operation.</returns>
-        private Task<object> GhGetToolAsync(JObject parameters)
-        {
-            // Parse filters
-            var attrFilters = parameters["attrFilters"]?.ToObject<List<string>>() ?? new List<string>();
-            var typeFilters = parameters["typeFilter"]?.ToObject<List<string>>() ?? new List<string>();
-            var objects = GHCanvasUtils.GetCurrentObjects();
-
-            // Filter by manual UI selection if provided
-            var selectedGuids = parameters["guidFilter"]?.ToObject<List<string>>() ?? new List<string>();
-            if (selectedGuids.Any())
-            {
-                var set = new HashSet<string>(selectedGuids);
-                objects = objects.Where(o => set.Contains(o.InstanceGuid.ToString())).ToList();
-            }
-
-            var connectionDepth = parameters["connectionDepth"]?.ToObject<int>() ?? 0;
-            var (includeTypes, excludeTypes) = ParseIncludeExclude(typeFilters, TypeSynonyms);
-            var (includeTags, excludeTags) = ParseIncludeExclude(attrFilters, FilterSynonyms);
-
-            // Apply typeFilters on base objects
-            var typeFiltered = new List<IGH_ActiveObject>(objects);
-            if (includeTypes.Any())
-            {
-                var tf = new List<IGH_ActiveObject>();
-                if (includeTypes.Contains("PARAMS"))
-                {
-                    tf.AddRange(objects.OfType<IGH_Param>());
-                }
-
-                if (includeTypes.Contains("COMPONENTS"))
-                {
-                    tf.AddRange(objects.OfType<GH_Component>());
-                }
-
-                if (includeTypes.Overlaps(new[] { "INPUT", "OUTPUT", "PROCESSING", "ISOLATED" }))
-                {
-                    var tempDoc = GHDocumentUtils.GetObjectsDetails(objects);
-                    var incd = new Dictionary<Guid, int>();
-                    var outd = new Dictionary<Guid, int>();
-                    foreach (var conn in tempDoc.Connections)
-                    {
-                        outd[conn.From.ComponentId] = (outd.TryGetValue(conn.From.ComponentId, out var ov) ? ov : 0) + 1;
-                        incd[conn.To.ComponentId] = (incd.TryGetValue(conn.To.ComponentId, out var iv) ? iv : 0) + 1;
-                    }
-
-                    if (includeTypes.Contains("INPUT"))
-                    {
-                        tf.AddRange(objects.Where(c => !incd.ContainsKey(c.InstanceGuid) && outd.ContainsKey(c.InstanceGuid)));
-                    }
-
-                    if (includeTypes.Contains("OUTPUT"))
-                    {
-                        tf.AddRange(objects.Where(c => !outd.ContainsKey(c.InstanceGuid) && incd.ContainsKey(c.InstanceGuid)));
-                    }
-
-                    if (includeTypes.Contains("PROCESSING"))
-                    {
-                        tf.AddRange(objects.Where(c => incd.ContainsKey(c.InstanceGuid) && outd.ContainsKey(c.InstanceGuid)));
-                    }
-
-                    if (includeTypes.Contains("ISOLATED"))
-                    {
-                        tf.AddRange(objects.Where(c => !incd.ContainsKey(c.InstanceGuid) && !outd.ContainsKey(c.InstanceGuid)));
-                    }
-                }
-
-                typeFiltered = tf.ToList();
-            }
-
-            if (excludeTypes.Any())
-            {
-                if (!includeTypes.Any())
-                {
-                    typeFiltered = new List<IGH_ActiveObject>(objects);
-                }
-
-                if (excludeTypes.Contains("PARAMS"))
-                {
-                    typeFiltered.RemoveAll(o => o is IGH_Param);
-                }
-
-                if (excludeTypes.Contains("COMPONENTS"))
-                {
-                    typeFiltered.RemoveAll(o => o is GH_Component);
-                }
-
-                if (excludeTypes.Overlaps(new[] { "INPUT", "OUTPUT", "PROCESSING", "ISOLATED" }))
-                {
-                    var tempDoc = GHDocumentUtils.GetObjectsDetails(typeFiltered);
-                    var incd = new Dictionary<Guid, int>();
-                    var outd = new Dictionary<Guid, int>();
-                    foreach (var conn in tempDoc.Connections)
-                    {
-                        outd[conn.From.ComponentId] = (outd.TryGetValue(conn.From.ComponentId, out var ov) ? ov : 0) + 1;
-                        incd[conn.To.ComponentId] = (incd.TryGetValue(conn.To.ComponentId, out var iv) ? iv : 0) + 1;
-                    }
-
-                    if (excludeTypes.Contains("INPUT"))
-                    {
-                        typeFiltered.RemoveAll(c => !incd.ContainsKey(c.InstanceGuid) && outd.ContainsKey(c.InstanceGuid));
-                    }
-
-                    if (excludeTypes.Contains("OUTPUT"))
-                    {
-                        typeFiltered.RemoveAll(c => !outd.ContainsKey(c.InstanceGuid) && incd.ContainsKey(c.InstanceGuid));
-                    }
-
-                    if (excludeTypes.Contains("PROCESSING"))
-                    {
-                        typeFiltered.RemoveAll(c => incd.ContainsKey(c.InstanceGuid) && outd.ContainsKey(c.InstanceGuid));
-                    }
-
-                    if (excludeTypes.Contains("ISOLATED"))
-                    {
-                        typeFiltered.RemoveAll(c => !incd.ContainsKey(c.InstanceGuid) && !outd.ContainsKey(c.InstanceGuid));
-                    }
-                }
-            }
-
-            objects = typeFiltered;
-
-            // Apply includes
-            List<IGH_ActiveObject> resultObjects;
-            if (includeTags.Any())
-            {
-                resultObjects = new List<IGH_ActiveObject>();
-                if (includeTags.Contains("SELECTED"))
-                {
-                    resultObjects.AddRange(objects.Where(o => o.Attributes.Selected));
-                }
-
-                if (includeTags.Contains("UNSELECTED"))
-                {
-                    resultObjects.AddRange(objects.Where(o => !o.Attributes.Selected));
-                }
-
-                if (includeTags.Contains("ENABLED"))
-                {
-                    resultObjects.AddRange(objects.OfType<GH_Component>().Where(c => !c.Locked).Cast<IGH_ActiveObject>());
-                }
-
-                if (includeTags.Contains("DISABLED"))
-                {
-                    resultObjects.AddRange(objects.OfType<GH_Component>().Where(c => c.Locked).Cast<IGH_ActiveObject>());
-                }
-
-                if (includeTags.Contains("ERROR"))
-                {
-                    resultObjects.AddRange(objects.Where(o => o.RuntimeMessages(GH_RuntimeMessageLevel.Error).Any()));
-                }
-
-                if (includeTags.Contains("WARNING"))
-                {
-                    resultObjects.AddRange(objects.Where(o => o.RuntimeMessages(GH_RuntimeMessageLevel.Warning).Any()));
-                }
-
-                if (includeTags.Contains("REMARK"))
-                {
-                    resultObjects.AddRange(objects.Where(o => o.RuntimeMessages(GH_RuntimeMessageLevel.Remark).Any()));
-                }
-
-                if (includeTags.Contains("PREVIEWCAPABLE"))
-                {
-                    resultObjects.AddRange(objects.OfType<GH_Component>().Where(c => c.IsPreviewCapable).Cast<IGH_ActiveObject>());
-                }
-
-                if (includeTags.Contains("NOTPREVIEWCAPABLE"))
-                {
-                    resultObjects.AddRange(objects.OfType<GH_Component>().Where(c => !c.IsPreviewCapable).Cast<IGH_ActiveObject>());
-                }
-
-                if (includeTags.Contains("PREVIEWON"))
-                {
-                    resultObjects.AddRange(objects.OfType<GH_Component>().Where(c => c.IsPreviewCapable && !c.Hidden).Cast<IGH_ActiveObject>());
-                }
-
-                if (includeTags.Contains("PREVIEWOFF"))
-                {
-                    resultObjects.AddRange(objects.OfType<GH_Component>().Where(c => c.IsPreviewCapable && c.Hidden).Cast<IGH_ActiveObject>());
-                }
-            }
-            else
-            {
-                resultObjects = new List<IGH_ActiveObject>(objects);
-            }
-
-            // Apply excludes
-            if (excludeTags.Contains("SELECTED"))
-            {
-                resultObjects.RemoveAll(o => o.Attributes.Selected);
-            }
-
-            if (excludeTags.Contains("UNSELECTED"))
-            {
-                resultObjects.RemoveAll(o => !o.Attributes.Selected);
-            }
-
-            if (excludeTags.Contains("ENABLED"))
-            {
-                resultObjects.RemoveAll(o => o is GH_Component c && !c.Locked);
-            }
-
-            if (excludeTags.Contains("DISABLED"))
-            {
-                resultObjects.RemoveAll(o => o is GH_Component c && c.Locked);
-            }
-
-            if (excludeTags.Contains("ERROR"))
-            {
-                resultObjects.RemoveAll(o => o.RuntimeMessages(GH_RuntimeMessageLevel.Error).Any());
-            }
-
-            if (excludeTags.Contains("WARNING"))
-            {
-                resultObjects.RemoveAll(o => o.RuntimeMessages(GH_RuntimeMessageLevel.Warning).Any());
-            }
-
-            if (excludeTags.Contains("REMARK"))
-            {
-                resultObjects.RemoveAll(o => o.RuntimeMessages(GH_RuntimeMessageLevel.Remark).Any());
-            }
-
-            if (excludeTags.Contains("PREVIEWCAPABLE"))
-            {
-                resultObjects.RemoveAll(o => o is GH_Component c && c.IsPreviewCapable);
-            }
-
-            if (excludeTags.Contains("NOTPREVIEWCAPABLE"))
-            {
-                resultObjects.RemoveAll(o => o is GH_Component c && !c.IsPreviewCapable);
-            }
-
-            if (excludeTags.Contains("PREVIEWON"))
-            {
-                resultObjects.RemoveAll(o => o is GH_Component c && c.IsPreviewCapable && !c.Hidden);
-            }
-
-            if (excludeTags.Contains("PREVIEWOFF"))
-            {
-                resultObjects.RemoveAll(o => o is GH_Component c && c.IsPreviewCapable && c.Hidden);
-            }
-
-            if (connectionDepth > 0)
-            {
-                var allObjects = GHCanvasUtils.GetCurrentObjects();
-                var fullDoc = GHDocumentUtils.GetObjectsDetails(allObjects);
-                var edges = fullDoc.Connections.Select(c => (c.From.ComponentId, c.To.ComponentId));
-                var initialIds = resultObjects.Select(o => o.InstanceGuid);
-                var expandedIds = ConnectionGraphUtils.ExpandByDepth(edges, initialIds, connectionDepth);
-                var idMap = allObjects.ToDictionary(o => o.InstanceGuid, o => o);
-                resultObjects = expandedIds
-                    .Select(g => idMap.TryGetValue(g, out var obj) ? obj : null)
-                    .Where(o => o != null)
-                    .ToList();
-            }
-
-            var document = GHDocumentUtils.GetObjectsDetails(resultObjects);
-
-            // only keep connections where both components are in our filtered set
-            var allowed = resultObjects.Select(o => o.InstanceGuid).ToHashSet();
-            document.Connections = document.Connections
-                .Where(c => allowed.Contains(c.From.ComponentId)
-                         && allowed.Contains(c.To.ComponentId))
-                .ToList();
-
-            // Get names and guids
-            var names = document.Components.Select(c => c.Name).Distinct().ToList();
-            var guids = document.Components.Select(c => c.InstanceGuid.ToString()).Distinct().ToList();
-
-            // Serialize document
-            var json = JsonConvert.SerializeObject(document, Formatting.None);
-
-            // Package result with classifications
-            var result = new JObject
-            {
-                ["names"] = JArray.FromObject(names),
-                ["guids"] = JArray.FromObject(guids),
-                ["json"] = json,
-            };
-
-            return Task.FromResult<object>(result);
-        }
-        #endregion
-
-        #region GhRetrieve
-
-        /// <summary>
-        /// Executes the Grasshopper list component types tool.
-        /// </summary>
-        private Task<object> GhRetrieveToolAsync(JObject parameters)
-        {
-            var server = Instances.ComponentServer;
-            var categoryFilters = parameters["categoryFilter"]?.ToObject<List<string>>() ?? new List<string>();
-            var (includeCats, excludeCats) = ParseIncludeExclude(categoryFilters, CategorySynonyms);
-
-            // Retrieve all component proxies in one call
-            var proxies = server.ObjectProxies.ToList();
-
-            // Apply include filters
-            if (includeCats.Any())
-            {
-                proxies = proxies.Where(p => p.Desc.Category != null && includeCats.Contains(p.Desc.Category.ToUpperInvariant())).ToList();
-            }
-
-            // Apply exclude filters
-            if (excludeCats.Any())
-            {
-                proxies = proxies.Where(p => p.Desc.Category == null || !excludeCats.Contains(p.Desc.Category.ToUpperInvariant())).ToList();
-            }
-
-            var list = proxies.Select(p =>
-            {
-                var instance = GHObjectFactory.CreateInstance(p);
-                List<object> inputs;
-                List<object> outputs;
-                if (instance is IGH_Component comp)
-                {
-                    inputs = GHParameterUtils.GetAllInputs(comp)
-                        .Select(param => new
-                        {
-                            name = param.Name,
-                            description = param.Description,
-                            dataType = param.GetType().Name,
-                            access = param.Access.ToString(),
-                        })
-                        .Cast<object>()
-                        .ToList();
-                    outputs = GHParameterUtils.GetAllOutputs(comp)
-                        .Select(param => new
-                        {
-                            name = param.Name,
-                            description = param.Description,
-                            dataType = param.GetType().Name,
-                            access = param.Access.ToString(),
-                        })
-                        .Cast<object>()
-                        .ToList();
-                }
-                else if (instance is IGH_Param param)
-                {
-                    inputs = new List<object>();
-                    outputs = new List<object>
-                    {
-                        new
-                        {
-                            name = param.Name,
-                            description = param.Description,
-                            dataType = param.GetType().Name,
-                            access = param.Access.ToString(),
-                        },
-                    };
-                }
-                else
-                {
-                    inputs = new List<object>();
-                    outputs = new List<object>();
-                }
-
-                return new
-                {
-                    name = p.Desc.Name,
-                    nickname = p.Desc.NickName,
-                    category = p.Desc.Category,
-                    subCategory = p.Desc.SubCategory,
-                    guid = p.Guid.ToString(),
-                    description = p.Desc.Description,
-                    keywords = p.Desc.Keywords,
-                    inputs,
-                    outputs,
-                };
-            }).ToList();
-            var names = list.Select(x => x.name).Distinct().ToList();
-            var guids = list.Select(x => x.guid).Distinct().ToList();
-            var json = JsonConvert.SerializeObject(list, Formatting.None);
-            var result = new JObject
-            {
-                ["count"] = list.Count,
-                ["names"] = JArray.FromObject(names),
-                ["guids"] = JArray.FromObject(guids),
-                ["json"] = json,
-            };
-            return Task.FromResult<object>(result);
-        }
-        #endregion
-
-        #region GhCategories
-
-        /// <summary>
-        /// Executes the Grasshopper categories listing tool with optional soft string filter.
-        /// </summary>
-        private Task<object> GhCategoriesToolAsync(JObject parameters)
-        {
-            var server = Instances.ComponentServer;
-            var filterString = parameters["filter"]?.ToObject<string>() ?? string.Empty;
-            var tokens = Regex.Replace(filterString, @"[,;\-_]", " ")
-                .Split(new[] { ' ' }, StringSplitOptions.RemoveEmptyEntries)
-                .Select(t => t.ToLowerInvariant())
-                .ToList();
-
-            var proxies = server.ObjectProxies;
-            var dict = new Dictionary<string, HashSet<string>>(StringComparer.OrdinalIgnoreCase);
-            foreach (var p in proxies)
-            {
-                var cat = p.Desc.Category ?? string.Empty;
-                var sub = p.Desc.SubCategory ?? string.Empty;
-                if (!dict.TryGetValue(cat, out var subs))
-                {
-                    subs = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
-                    dict[cat] = subs;
-                }
-
-                if (!string.IsNullOrEmpty(sub))
-                {
-                    subs.Add(sub);
-                }
-            }
-
-            var result = new List<object>();
-            foreach (var kv in dict)
-            {
-                var cat = kv.Key;
-                var subs = kv.Value;
-                if (!tokens.Any())
-                {
-                    result.Add(new { category = cat, subCategories = subs.ToList() });
-                }
-                else
-                {
-                    bool catMatch = tokens.Any(tok => cat.ToLowerInvariant().Contains(tok));
-                    var matchingSubs = subs.Where(s => tokens.Any(tok => s.ToLowerInvariant().Contains(tok))).ToList();
-                    if (catMatch)
-                    {
-                        result.Add(new { category = cat, subCategories = subs.ToList() });
-                    }
-                    else if (matchingSubs.Any())
-                    {
-                        result.Add(new { category = cat, subCategories = matchingSubs });
-                    }
-                }
-            }
-
-            var jResult = new JObject
-            {
-                ["count"] = result.Count,
-                ["categories"] = JArray.FromObject(result),
-            };
-            return Task.FromResult<object>(jResult);
-        }
-        #endregion
     }
 }
