+++ src/SmartHopper.Core.Grasshopper/Tools/WebTools.cs
@@ -8,36 +8,68 @@
  * version 3 of the License, or (at your option) any later version.
  */
 
-/*
- * WebTools.cs
- * Defines AI tool for fetching webpage text content, stripping HTML and respecting robots.txt.
- */
-
 using System;
 using System.Collections.Generic;
-using System.Diagnostics;
-using System.Linq;
-using System.Net.Http;
-using System.Text.RegularExpressions;
-using System.Threading.Tasks;
-using HtmlAgilityPack;
-using Newtonsoft.Json.Linq;
-using SmartHopper.Config.Interfaces;
-using SmartHopper.Config.Models;
 
 namespace SmartHopper.Core.Grasshopper.Tools
 {
+    /// <summary>
+    /// Provides web-related utility functions, including robots.txt parsing.
+    /// </summary>
     public class WebTools
     {
-        private readonly List<string> disallowed = new();
+        private readonly List<string> _disallowed = new();
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="RobotsTxtParser"/> class.
-        /// Simple robots.txt parser supporting User-agent: * and Disallow directives.
+        /// Initializes a new instance of the <see cref="WebTools"/> class.
         /// </summary>
-        public RobotsTxtParser(string content)
+        /// <param name="robotsTxtContent">The content of the robots.txt file to parse.</param>
+        public WebTools(string robotsTxtContent = null)
         {
+            if (!string.IsNullOrEmpty(robotsTxtContent))
+            {
+                this.ParseRobotsTxt(robotsTxtContent);
+            }
+        }
+
+        /// <summary>
+        /// Checks if a given path is allowed by the robots.txt rules.
+        /// </summary>
+        /// <param name="path">The path to check (e.g., "/some/path")</param>
+        /// <returns>True if the path is allowed, false if disallowed by robots.txt</returns>
+        public bool IsPathAllowed(string path)
+        {
+            if (string.IsNullOrEmpty(path))
+            {
+                return true;
+            }
+
+            // Ensure path starts with /
+            string normalizedPath = path.StartsWith('/') ? path : "/" + path;
+
+            // Check against all disallowed patterns
+            foreach (var pattern in _disallowed)
+            {
+                if (string.IsNullOrEmpty(pattern))
+                {
+                    continue;
+                }
+
+                // Simple pattern matching - could be enhanced with regex for full robots.txt spec
+                if (normalizedPath.StartsWith(pattern.TrimEnd('*'), StringComparison.OrdinalIgnoreCase))
+                {
+                    return false;
+                }
+            }
+
+            return true;
+        }
+
+        private void ParseRobotsTxt(string content)
+        {
+            this._disallowed.Clear();
             bool appliesToAll = false;
+
             foreach (var line in content.Split('\n'))
             {
                 string trimmed = line.Trim();
@@ -58,36 +90,11 @@ namespace SmartHopper.Core.Grasshopper.Tools
                 {
                     appliesToAll = value == "*";
                 }
-                else if (field == "disallow" && appliesToAll)
+                else if (field == "disallow" && appliesToAll && !string.IsNullOrEmpty(value))
                 {
-                    this.disallowed.Add(value);
-                }
-                else if (field == "allow" && appliesToAll)
-                {
-                    // Could implement allow rules, but ignored for simplicity
+                    _disallowed.Add(value);
                 }
             }
-        }
-
-        /// <summary>
-        /// Returns true if the given path is allowed to be fetched.
-        /// </summary>
-        public bool IsAllowed(string path)
-        {
-            foreach (var rule in this.disallowed)
-            {
-                if (string.IsNullOrEmpty(rule))
-                {
-                    continue;
-                }
-
-                if (path.StartsWith(rule))
-                {
-                    return false;
-                }
-            }
-
-            return true;
         }
     }
 }
